windows: Add Git Bash to the path
The libcxx tests currently require bash to be available (although
that might change), and the bash provided with Git is enough for
fulfilling this need. (Bash can also be useful for running other
scripts in the libcxx CI chain outside of the tests themselves.)

For the main LLVM testing, this works once
https://reviews.llvm.org/D98858 is merged.

This also allows getting rid of GnuWin altogether.