Adjust repliace count for windows buildkite cluster
The replica count was originally adjusted to 15 when some additional
nodes were added to the cluster. However, there are only 15 machines,
which means that either a buildkite pod or the post-commit buildbot pod
will not schedule. In addition, we are running into GCP quota limits
where we only actually have 14 nodes, so drop the number of buildkite
replicas down to 13 for now so we can schedule the buildbot pod.